requirements.txt
# Travel Assistant - Production Dependencies
# ============================================
# AI-powered travel recommendation system with memory and caching

# Core FastAPI Framework & ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI & Machine Learning
google-generativeai==0.8.3
mem0ai==1.0.1
sentence-transformers==5.1.2
scikit-learn==1.7.2
numpy==2.1.3
torch==2.9.1

# Caching & Database
redis==5.1.1
sqlalchemy==2.0.44

# HTTP & Networking
httpx==0.28.1
aiohttp==3.11.7
requests==2.32.5

# Data Processing & Validation
pydantic==2.12.4
pydantic-settings==2.6.1

# Environment & Configuration
python-dotenv==1.0.1
python-multipart==0.0.12

# Workflow & Graph Processing
langgraph==1.0.1
langgraph-checkpoint==2.1.1
langchain-core==0.3.80
langchain-google-genai==2.0.5

# Security & Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==44.0.0

# Logging & Monitoring
loguru==0.7.3
structlog==24.4.0

# Testing & Development
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx-sse==0.4.0

# Utilities
typing-extensions==4.15.0
python-dateutil==2.9.0
pytz==2025.2
aiofiles==24.1.0

# Production Deployment
gunicorn==23.0.0

# Optional: Enhanced Performance
uvloop==0.25.0

# Optional: Monitoring & Observability
prometheus-client==0.21.1
sentry-sdk[fastapi]==2.19.2